Going to use ArUco Marker Tracking
Basically, I'll have 2 markers, one on top of the egg, and the other on top of the wand. And a webcam will be suspended looking down at the egg. ArUco Library for Open CV can detect the positions of of each marker and I can use that data and do some math on it to translate to camera position and rotation